Move legacy & update to beta libs
