+++ package.json
@@ -20,15 +20,10 @@
 		".": {
 			"import": "./dist/index.js",
 			"default": "./dist/index.cjs"
-		},
-		"./infinite": {
-			"import": "./dist/infinite.js",
-			"default": "./dist/infinite.cjs"
 		}
 	},
 	"files": [
-		"dist",
-		"infinite"
+		"dist"
 	],
 	"scripts": {
 		"build": "tsup",
@@ -42,34 +37,34 @@
 		"test": "vitest"
 	},
 	"peerDependencies": {
-		"@trpc/client": ">=9.23.2",
-		"@trpc/server": ">=9.23.2",
+		"@trpc/client": "10.0.0-proxy-beta.14",
+		"@trpc/server": "10.0.0-proxy-beta.14",
 		"react": "^17 || ^18",
-		"swr": "^1"
+		"swr": "^2.0.0-beta.7"
 	},
 	"devDependencies": {
 		"@sachinraja/eslint-config": "0.1.1",
-		"@testing-library/dom": "8.16.0",
-		"@testing-library/jest-dom": "5.16.4",
-		"@testing-library/react": "13.3.0",
-		"@testing-library/user-event": "14.3.0",
-		"@trpc/client": "9.26.2",
-		"@trpc/server": "9.26.2",
-		"@types/node": "18.6.3",
-		"@types/react": "18.0.15",
-		"@vitejs/plugin-react": "2.0.0",
-		"dprint": "0.30.3",
-		"eslint": "8.20.0",
+		"@testing-library/dom": "8.18.1",
+		"@testing-library/jest-dom": "5.16.5",
+		"@testing-library/react": "13.4.0",
+		"@testing-library/user-event": "14.4.3",
+		"@trpc/client": "10.0.0-proxy-beta.14",
+		"@trpc/server": "10.0.0-proxy-beta.14",
+		"@types/node": "18.8.2",
+		"@types/react": "18.0.21",
+		"@vitejs/plugin-react": "2.1.0",
+		"dprint": "0.32.1",
+		"eslint": "8.24.0",
 		"get-port": "6.1.2",
 		"husky": "8.0.1",
-		"jsdom": "20.0.0",
+		"jsdom": "20.0.1",
 		"nano-staged": "0.8.0",
 		"npm-run-all": "4.1.5",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
-		"swr": "1.3.0",
-		"tsup": "6.2.1",
-		"typescript": "4.7.4",
+		"swr": "^2.0.0-beta.7",
+		"tsup": "6.2.3",
+		"typescript": "4.8.4",
 		"undici": "5.8.0",
 		"vite": "3.0.4",
 		"vitest": "0.20.2",
