+++ playground/package.json
@@ -14,7 +14,7 @@
 		"zod": "3.19.1"
 	},
 	"devDependencies": {
-		"@types/node": "18.6.3",
+		"@types/node": "18.8.2",
 		"@types/react-dom": "18.0.6"
 	}
 }
